# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "panda-wiki API documentation",
        "title": "panda-wiki API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/knowledge_base/user/delete": {
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Remove user from knowledge base",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge_base"
                ],
                "summary": "KBUserDelete",
                "parameters": [
                    {
                        "type": "string",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/knowledge_base/user/invite": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Invite user to knowledge base",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge_base"
                ],
                "summary": "KBUserInvite",
                "parameters": [
                    {
                        "description": "Invite User Request",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.KBUserInviteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/knowledge_base/user/list": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "KBUserList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge_base"
                ],
                "summary": "KBUserList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Knowledge Base ID",
                        "name": "kb_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PWResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/v1.KBUserListItemResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/knowledge_base/user/update": {
            "patch": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update user permission in knowledge base",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge_base"
                ],
                "summary": "KBUserUpdate",
                "parameters": [
                    {
                        "description": "Update User Permission Request",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.KBUserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "consts.UserKBPermission": {
            "type": "string",
            "enum": [
                "",
                "full_control",
                "doc_manage",
                "data_operate"
            ],
            "x-enum-comments": {
                "UserKBPermissionDataOperate": "数据运营",
                "UserKBPermissionDocManage": "文档管理",
                "UserKBPermissionFullControl": "完全控制",
                "UserKBPermissionNull": "无权限"
            },
            "x-enum-descriptions": [
                "无权限",
                "完全控制",
                "文档管理",
                "数据运营"
            ],
            "x-enum-varnames": [
                "UserKBPermissionNull",
                "UserKBPermissionFullControl",
                "UserKBPermissionDocManage",
                "UserKBPermissionDataOperate"
            ]
        },
        "consts.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-comments": {
                "UserRoleAdmin": "管理员",
                "UserRoleUser": "普通用户"
            },
            "x-enum-descriptions": [
                "管理员",
                "普通用户"
            ],
            "x-enum-varnames": [
                "UserRoleAdmin",
                "UserRoleUser"
            ]
        },
...
        "domain.PWResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "v1.KBUserInviteReq": {
            "type": "object",
            "required": [
                "kb_id",
                "perm",
                "user_id"
            ],
            "properties": {
                "kb_id": {
                    "type": "string"
                },
                "perm": {
                    "enum": [
                        "full_control",
                        "doc_manage",
                        "data_operate"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.UserKBPermission"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "v1.KBUserListItemResp": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "perms": {
                    "$ref": "#/definitions/consts.UserKBPermission"
                },
                "role": {
                    "$ref": "#/definitions/consts.UserRole"
                }
            }
        },
        "v1.KBUserUpdateReq": {
            "type": "object",
            "required": [
                "kb_id",
                "perm",
                "user_id"
            ],
            "properties": {
                "kb_id": {
                    "type": "string"
                },
                "perm": {
                    "enum": [
                        "full_control",
                        "doc_manage",
                        "data_operate"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/consts.UserKBPermission"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearerAuth": {
            "description": "Type \"Bearer\" + a space + your token to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}